command: consolidate interfaces + change methods
- move the `Command` interface into `command.go`
- remove the `HelpFlag` interface
- change `Help` methods to `HelpRequested`
- various minor doc changes
