+++ internal/command/help.go
@@ -2,20 +2,12 @@ package command
 
 import "flag"
 
-type (
-	// HelpArg implements [HelpFlag].
-	HelpArg bool
+// HelpArg implements [HelpFlag].
+type HelpArg bool
 
-	// HelpFlag's `Help` method will signify
-	// whether a caller requested help or not.
-	HelpFlag interface {
-		Help() bool
-	}
-)
-
-// Help will return true if the help flag
-// was present when parsing.
-func (b HelpArg) Help() bool { return bool(b) }
+// HelpRequested will return true if a help flag
+// was set during parsing.
+func (b HelpArg) HelpRequested() bool { return bool(b) }
 
 // BindFlags defines a `-help` flag in the [flag.FlagSet].
 func (b *HelpArg) BindFlags(fs *flag.FlagSet) {
